Мигрировать проект на TS

+++ service_books/src/classes/clBooksRepository.ts
@@ -1,28 +1,30 @@
-//const {books} = im('../storage/storage2.ts')
-
-//import { storage.books as books } from '../storage/storage2'
 import { injectable } from 'inversify'
 import { ifBook } from './ifBook'
-import { cBook } from './cBook2'
-import fs from 'fs'
-import axios from 'axios'
+import { clBook } from './clBook'
 import Books from '../models/books'
-import Comments from '../models/comments'
+import fs from 'fs'
 
 @injectable()
 export abstract class clBooksRepository{
+    // количество книг в коллекции
+    async get_count():Promise<number>{
+        return await Books.countDocuments()       
+    }
     // создание книги
-    async createBook(book: ifBook, files: any):Promise<ifBook> {        
+    async createBook(book: ifBook, files: any = null):Promise<ifBook> {        
 
-        let fileCover = ""
-        if (files.cover !== undefined)
-            fileCover = files.cover[0].path
-    
-        let fileName = ""
-        let fileBook = ""
-        if (files.book !== undefined) {
-            fileName = files.book[0].filename
-            fileBook = files.book[0].path
+        let fileCover = book.fileCover
+        let fileName  = book.fileName
+        let fileBook  = book.fileBook
+
+        if (files) {            
+            if (files.cover !== undefined)
+                fileCover = files.cover[0].path       
+
+            if (files.book !== undefined) {
+                fileName = files.book[0].filename
+                fileBook = files.book[0].path
+            }
         }
     
         const newBook = new Books({
@@ -38,6 +40,7 @@ export abstract class clBooksRepository{
             await newBook.save()
             return book
         } catch (e) {
+            console.log(e)
             return null
         }        
     }
@@ -46,8 +49,8 @@ export abstract class clBooksRepository{
 
         try {
             const dbBook = await Books.findById(id).select('-__v')
-            const {title, description, authors, favorite, fileCover, fileName, fileBook} = dbBook
-            const book = new cBook (title, description, authors, favorite, fileCover, fileName, fileBook)
+            const {title, description, authors, favorite, fileCover, fileName, fileBook, id: dbId } = dbBook
+            const book = new clBook (title, description, authors, favorite, fileCover, fileName, fileBook, dbId)
             return book
         } catch (e) { 
             return null
@@ -58,11 +61,11 @@ export abstract class clBooksRepository{
         try {
             const dbBooks = await Books.find().select('-__v')
             
-            let books: cBook[]
+            let books: clBook[]
             let i: number
             for (i = 1; i < dbBooks.length; i++ ){
-                const {title, description, authors, favorite, fileCover, fileName, fileBook} = dbBooks[i]
-                const book = new cBook (title, description, authors, favorite, fileCover, fileName, fileBook)   
+                const {title, description, authors, favorite, fileCover, fileName, fileBook, id} = dbBooks[i]
+                const book = new clBook (title, description, authors, favorite, fileCover, fileName, fileBook, id)   
                 books.push(book)          
             }
             return books;
@@ -99,20 +102,21 @@ export abstract class clBooksRepository{
     
             let fileName = dbBook.fileName
             let fileBook = dbBook.fileBook
-            if (files.book !== undefined) {
-                // Пришел новый файл книги, надо удалить старый файл
-                console.log(`Нужно удалить файл книги - ${fileBook}`)
-                try{
-                    fs.unlinkSync(fileBook)
-                } catch (e) {
-                    console.log(`Файл ${fileBook} не удален`)
-                    console.log(e)
-                }        
-                fileName = files.book[0].filename
-                fileBook = files.book[0].path
-            }
+            if (files)
+                if (files.book !== undefined) {
+                    // Пришел новый файл книги, надо удалить старый файл
+                    console.log(`Нужно удалить файл книги - ${fileBook}`)
+                    try{
+                        fs.unlinkSync(fileBook)
+                    } catch (e) {
+                        console.log(`Файл ${fileBook} не удален`)
+                        console.log(e)
+                    }        
+                    fileName = files.book[0].filename
+                    fileBook = files.book[0].path
+                }
     
-            const book = new cBook (title, description, authors, favorite, fileCover, fileName, fileBook)
+            const book = new clBook (title, description, authors, favorite, fileCover, fileName, fileBook, id)
             try {
                 await Books.findByIdAndUpdate(id, {
                     title,
@@ -151,8 +155,8 @@ export abstract class clBooksRepository{
                 console.log(e)
             }
 
-            const {title, description, authors, favorite, fileCover, fileName, fileBook} = dbBook
-            const book = new cBook (title, description, authors, favorite, fileCover, fileName, fileBook)
+            const {title, description, authors, favorite, fileCover, fileName, fileBook, id: dbId} = dbBook
+            const book = new clBook (title, description, authors, favorite, fileCover, fileName, fileBook, dbId)
 
             await Books.deleteOne({_id: id})
